requirements.txt
altair==4.2.2
asttokens==2.2.1
attrs==23.1.0
autopep8==2.0.1
backcall==0.2.0
blinker==1.6.2
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
colorlover==0.3.0
comm==0.1.3
contourpy==1.0.7
cufflinks==0.17.3
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
distlib==0.3.6
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.17.0
filelock==3.8.0
fonttools==4.39.4
gitdb==4.0.10
GitPython==3.1.31
htmlmin==0.1.12
idna==3.4
ImageHash==4.3.1
importlib-metadata==6.6.0
ipykernel==6.23.1
ipython==8.13.2
ipywidgets==8.0.6
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mdurl==0.1.2
missingno==0.5.2
multimethod==1.9.1
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
numpy==1.23.5
packaging==23.1
pandas==1.5.3
pandas-profiling==3.6.6
parso==0.8.3
patsy==0.5.3
phik==0.12.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
plotly==5.14.1
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
pure-eval==0.2.2
pure-python-adb==0.3.0.dev0
pyarrow==12.0.0
pycodestyle==2.10.0
pydantic==1.10.7
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
pywin32==306
PyYAML==6.0
pyzmq==25.0.2
requests==2.28.2
rich==13.3.5
scipy==1.9.3
seaborn==0.12.2
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
statsmodels==0.13.5
streamlit==1.22.0
tangled-up-in-unicode==0.2.0
tenacity==8.2.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.2
tqdm==4.64.1
traitlets==5.9.0
typeguard==2.13.3
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.15
validators==0.20.0
virtualenv==20.17.0
visions==0.7.5
watchdog==3.0.0
wcwidth==0.2.6
widgetsnbextension==4.0.7
ydata-profiling==4.1.2
zipp==3.15.0
